pyproject.toml
[project]
name = "trade-mcp"
version = "0.1.0"
description = "Trading MCP project"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.10.1",
    "kiteconnect",
    "python-dotenv",
]